Prepare THREAD_TABLE ++++++++++++++++++++
As OpenSCAD does not support function pointers, it is not an extensible
approach to calculate the thread specs from pitch and Dpitch. Instead, I
pre-calculate *all* of the specs in a table. The idea is to extend this
table with further threads (such as metric threads, conic threads,
bottle threads, etc.) later on.

Step 1: Calculate THREAD_TABLE in Excel.